Moved Sqs2PollingClient to Sqs2Consumer as PollingTask
